Sign Hashids to prevent conflicts with regular IDs (#27)
This pull request encodes parameters with a key to prevent allow validation of hashid. The purpose is two prevent accidentally decoding an integer the happens to be a valid hashid. A check is done to ensure the param decodes to two integers and that the first integer matches the key.