encode nil IDs as nil
In "normal" rails, the `id` (and hence the value of `#to_param`) is nil
for records that have not yet been persisted.

Previously, including the `Hashid::Rails` concern broke this behaviour,
raising an exception when `Model.new.to_param` was called. (Cases where
you might want to get the ID of a new model include serializing it over
a JSON API where the "edit" endpoint returns a new model instance if
one does not already exist.)

This change handles the above case by checking for nil in
`model.hashid_encode`.