Allow configuring whether `to_param` is overridden
By default, hashid-rails overrides the `to_param` method on models to
return a hashid. This lets `link_to` and other Rails helpers "just
work". But there might be cases where you don't want this to happen, so
now you can turn it off in the config globally or on a per-model basis.