Add testing matrix for supported Rails versions
This change is a follow-up to the Ruby testing matrix. It introduces
Appraisal as a development dependency to run the test suite in parallel
against the five most recent versions of Rails. We might find over time
that we need to prune out some minor releases as the current matrix
runs 11 parallel builds, that is, 5 versions of Rails times 2 versions
of Ruby (2*5=10) then one additional run for Ruby 2.7 on Rails 6.0.